+++ back/domain/email/email.service.js
@@ -1,43 +1,47 @@
 var https = require('https');
-const querystring = require('querystring');
 var sgMail = require('./sendgrid');
 var mgMail = require('./mailgun');
-
-//var sgMail = require('@sendgrid/mail');
 
 var emailService = {
   sendEmail: sendEmail
 };
 
 function sendEmail(email) {
-  
-  var mail = sgMail.setMail(email);
-  var option = sgMail.setHeader();
-  
-  request(mail, option);
-  var mail01 = mgMail.setMail(email);
-  var option01 = mgMail.setHeader();
-  request(mail01, option01);
+
+  const makeRequest = async () => {
+
+    var result;
+    try {
+      result = await request(sgMail.setHeader(), sgMail.setMail(email));
+    } catch (e) {
+      result = await request(mgMail.setHeader(), mgMail.setMail(email));
+    }
+    return result;
+  }
+
+  return makeRequest();
+
 }
 
-function request(mail, options) {
-
-  const req = https.request(options, (res) => {
-    console.log('statusCode:', res.statusCode);
-    console.log('headers:', res.headers);
-  
-    res.on('data', (d) => {
-      process.stdout.write(d);
-    });
+function request(params, postData) {
+  return new Promise(function(resolve, reject) {
+      var req = https.request(params, function(res) {
+          // reject on bad status          
+          if (res.statusCode < 200 || res.statusCode >= 300) {
+              return reject(new Error('statusCode=' + res.statusCode));
+          } else {
+            resolve(res)
+          }
+      });
+      // reject on request error
+      req.on('error', function(err) {          
+          reject(err);
+      });
+      if (postData) {
+          req.write(postData);
+      }      
+      req.end();
   });
-  
-  req.on('error', (e) => {
-    console.error(e);
-  });
-  
-  req.write(mail);
-  req.end();
-
 }
 
 module.exports = emailService;